+++ moviemanager_backend/src/main/java/com/mtcompany/moviemanager/service/CategoryService.java
@@ -8,7 +8,7 @@ import java.util.List;
 public interface CategoryService {
     List<Category> findAll();
 
-    Category findById(int theId);
+    Category findById(Long theId);
 
     Category createNewCategory(Category category);
 }
